+++ aurelix/state.py
@@ -3,6 +3,7 @@ import fastapi
 import typing
 import databases
 import sqlalchemy as sa
+import jwt
 
 class DatabaseState(typing.TypedDict):
     engine: sa.engine.Engine
@@ -15,6 +16,7 @@ class AppState(typing.TypedDict):
     oidc_settings: schema.OIDCConfiguration
     models: dict[str, schema.ModelSpec]
     views: typing.Any # aurelix.crud.base.ExtensibleViewsApp
+    oidc_jwk_client: jwt.PyJWKClient
 
 APP_STATE: dict[fastapi.FastAPI, AppState] = {}